added tests for controllers

+++ src/main/java/app/service/impl/AttributeServiceImpl.java
@@ -2,17 +2,17 @@ package app.service.impl;
 
 import app.domain.Attribute;
 import app.domain.Point;
-import app.domain.Value;
+import app.exception.AttributeExistsException;
+import app.exception.EmptyListException;
 import app.exception.WrongIdException;
 import app.repository.AttributeRepository;
+import app.repository.PointRepository;
 import app.service.AttributeService;
 import lombok.AllArgsConstructor;
-import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -22,6 +22,7 @@ import java.util.Objects;
 public class AttributeServiceImpl implements AttributeService {
 
     private final AttributeRepository attributeRepository;
+    private final PointRepository pointRepository;
 
     @Override
     public Attribute getAttributeById(Long id) {
@@ -33,43 +34,38 @@ public class AttributeServiceImpl implements AttributeService {
     }
 
     @Override
-    public Attribute getAttributeByName(String name) {
-        return attributeRepository.findAttributeByName(name);
-    }
-
-    @Override
     public List<Attribute> getAllAttributes() {
-        return attributeRepository.findAll();
+        List<Attribute> attributes = attributeRepository.findAll();
+        if (attributes.isEmpty()) {
+            throw new EmptyListException();
+        }
+        return attributes;
     }
 
     @Override
     public void addNewAttribute(Attribute attribute) {
-        attributeRepository.save(attribute);
-    }
-
-    @Override
-    public void updateAttribute(Attribute attribute) {
+        Attribute attributeFromDb = attributeRepository.findAttributeByName(attribute.getName());
+        if (Objects.nonNull(attributeFromDb)) {
+            throw new AttributeExistsException();
+        }
         attributeRepository.save(attribute);
     }
 
     @Override
     public void deleteAttributeById(Long id) {
+        Attribute attribute = attributeRepository.findAttributeById(id);
+        if (Objects.isNull(attribute)) {
+            throw new WrongIdException();
+        }
         attributeRepository.deleteAttributeById(id);
     }
 
     @Override
-    public void deleteAttributeByName(String name) {
-        attributeRepository.deleteAttributeByName(name);
-    }
-
-    @Override
     public List<Point> getPointsWithAttribute(Long attributeId) {
-        Attribute attribute = attributeRepository.findAttributeById(attributeId);
-        Hibernate.initialize(attribute.getValues());
-        List<Value> values = attribute.getValues();
-        List<Point> points = new ArrayList<>();
-        for (Value v : values) {
-            points.add(v.getPoint());
+        Attribute attribute = getAttributeById(attributeId);
+        List<Point> points = pointRepository.findByValues_Attribute(attribute);
+        if (points.isEmpty()) {
+            throw new EmptyListException();
         }
         return points;
     }
